put dcss-related code into EpochSys
