+++ src/persist/EpochSys.cpp
@@ -1,5 +1,4 @@
 #include "EpochSys.hpp"
-#include "DCSS.hpp"
 
 #include <omp.h>
 #include <atomic>
@@ -118,6 +117,22 @@ namespace pds{
         //     }
         //     epoch_advancer->set_epoch_freq(env_epoch_advance);
         // }
+    }
+
+    void sc_desc_t::try_complete(EpochSys* esys, uint64_t addr){
+        nbptr_t _d = nbptr.load();
+        int ret = 0;
+        if(_d.val!=addr) return;
+        if(in_progress(_d)){
+            if(esys->check_epoch(cas_epoch)){
+                ret = 2;
+                ret |= commit(_d);
+            } else {
+                ret = 4;
+                ret |= abort(_d);
+            }
+        }
+        cleanup(_d);
     }
 
     bool EpochSys::check_epoch(uint64_t c){
